refactor: add on file images for prior year speakers
- Optimize assets using imageOptim
- Rename ChrisDemars photo to remove whitespace
