+++ src/data/2023/speakers.js
@@ -10,7 +10,7 @@ import BlakeLapum from '@/assets/images/speakers/Blake Lapum.png'
 import BrianMcKeiver from '@/assets/images/speakers/BrianMcKeiver.png'
 import CeclynnFielderSherrer from '@/assets/images/speakers/Ceclynn FielderSherrer.jpg'
 import ChadHamdan from '@/assets/images/speakers/Chad Hamdan.jpg'
-import ChrisDemars from '@/assets/images/speakers/Chris Demars.jpg'
+import ChrisDemars from '@/assets/images/speakers/ChrisDemars.jpg'
 import ChrisTurvey from '@/assets/images/speakers/Chris Turvey.jpg'
 import DanSchnau from '@/assets/images/speakers/Dan Schnau.png'
 import DanielHope from '@/assets/images/speakers/Daniel Hope.jpg'
