+++ src/data/2024/speakers.js
@@ -14,7 +14,7 @@ import JohnThompson from '@/assets/images/speakers/John Thompson.jpg'
 import SageMurakishi from '@/assets/images/speakers/Sage Murakishi.jpg'
 import GodfreyNolan from '@/assets/images/speakers/Godfrey Nolan.png'
 import DarioALencina from '@/assets/images/speakers/Dario A Lencina Talarico.jpeg'
-import ChrisDemars from '@/assets/images/speakers/Chris Demars.jpg'
+import ChrisDemars from '@/assets/images/speakers/ChrisDemars.jpg'
 import AliElSharif from '@/assets/images/speakers/Ali El-Sharif.jpeg'
 import Vikashagrawal from '@/assets/images/speakers/Vikash Agrawal.webp'
 import ScottStoll from '@/assets/images/speakers/Scott Stoll.png'
